Experiment with groundwork for the game backend
The goal of these early commits will be to construct a working MUD-like
game and get a solid understanding of the components necessary before
attempting to generalize it into an interface for common use.

In this update are a rudimentary input reader, preparations for error
handling, a rudimentary `Location` and `State` struct definition, etc.